Morgan-Stanley-Code-To-Give-Team-5-Track-A-main/backend/auth.py
"""
Verify Supabase Auth JWT and return the current user id.
Used by protected routes (create event, join/leave event).

Uses Supabase's JWKS endpoint (asymmetric ECC/RS256) so verification works
after JWT key rotation. No shared JWT secret required.
"""
import os
from typing import Annotated

import jwt
from dotenv import load_dotenv
from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
from jwt import PyJWKClient

load_dotenv()

SUPABASE_URL = os.environ.get("SUPABASE_URL")
if not SUPABASE_URL:
    raise RuntimeError(
        "Missing SUPABASE_URL in .env (Supabase Dashboard → Project Settings → API)"
    )
SUPABASE_URL = SUPABASE_URL.rstrip("/")
JWKS_URI = f"{SUPABASE_URL}/auth/v1/.well-known/jwks.json"

# Single client; PyJWKClient caches keys and refreshes when needed.
_jwks_client: PyJWKClient | None = None


def _get_jwks_client() -> PyJWKClient:
    global _jwks_client
    if _jwks_client is None:
        _jwks_client = PyJWKClient(JWKS_URI)
    return _jwks_client


security = HTTPBearer(auto_error=False)


def get_current_user_id(
    credentials: Annotated[HTTPAuthorizationCredentials | None, Depends(security)],
) -> str:
    if not credentials or credentials.credentials is None:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Not authenticated. Log in and send Authorization: Bearer <token>.",
        )
    token = credentials.credentials
    try:
        jwks_client = _get_jwks_client()
        signing_key = jwks_client.get_signing_key_from_jwt(token)
        # Supabase asymmetric keys: ES256 (ECC P-256) or RS256
        payload = jwt.decode(
            token,
            signing_key.key,
            algorithms=["ES256", "RS256"],
            audience="authenticated",
        )
    except jwt.ExpiredSignatureError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Token expired. Please log in again.",
        )
    except jwt.InvalidTokenError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid token. Please log in again.",
        )
    sub = payload.get("sub")
    if not sub:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid token: missing subject.",
        )
    return sub
